requirements.txt
discord.py
google-generativeai
flask~=3.1.2
protobuf~=5.29.5
python-dotenv~=1.2.1